requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
falcon==2.0.0
gunicorn==20.0.4
idna==2.10
murmurhash==1.0.2
numpy==1.19.0
pip==20.1.1
plac==1.1.3
preshed==3.0.2
PyYAML==5.3.1
requests==2.24.0
setuptools==46.4.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.48.0
ujson==3.0.0
urllib3==1.25.9
wasabi==0.7.1
wheel==0.34.2